+++ CHANGELOG.md
@@ -1,19 +1,38 @@
-
 # Change Log
+
 All notable changes to this project will be documented in this file.
+
+## [v0.3] - 2019-12-15
+
+### Added
+
+- ethbian-geth-admin.sh
+- ethbian-monitoring.sh
+- geth_peers_geo2influx.py
+- grafana worldmap plugin
+
+### Removed
+
+- ethbian-geth-upgrade.sh
+
+### Changed
+
+- ethbian_install.sh
+- ethbian_install.test.sh
 
 ## [v0.2] - 2019-12-05
 
 ### Added
+
 - influxdb
 - collectd
 - grafana
 - ethbian-geth-upgrade.sh
 
 ### Removed
+
 - webfs
- 
+
 ## [v0.1] - 2019-11-17
 
-* initial release
- 
+- initial release